refactor: extract common functions to mod.rs
